+++ kicost/edas/eda_kicad.py
@@ -75,10 +75,9 @@ def get_part_groups(in_file, ignore_fields, variant):
                 elif SEPRTR not in name: # No separator, so get global field value.
                     name = field_name_translations.get(name, name)
                     value = str(f.string)
-                    if value and not name in fields:
-                        # Do not create empty fields. This is usefull
-                        # when used more than one `manf#` alias in one designator.
-                        fields[name] = value 
+                    if value:
+                        fields[name] = value # Do not create empty fields. This is useful
+                                             # when used more than one `manf#` alias in one designator.
                 else:
                     # Now look for fields that start with 'kicost' and possibly
                     # another dot-separated variant field and store their values.